clean up e2e tests
