+++ apps/web/playwright/insights.e2e.ts
@@ -237,4 +237,39 @@ test.describe("Insights", async () => {
     // Wait for the download process to complete and save the downloaded file somewhere.
     await download.saveAs("./" + "test-insights.csv");
   });
+
+  test("should render all ChartCard components with expected titles", async ({ page, users }) => {
+    const owner = await users.create();
+    const member = await users.create();
+
+    await createTeamsAndMembership(owner.id, member.id);
+
+    await owner.apiLogin();
+
+    await page.goto("/insights");
+
+    const expectedChartTitles = [
+      "Events",
+      "Performance",
+      "Event trends",
+      "Bookings by hour",
+      "Average event duration",
+      "Recent no-show guests",
+      "Most bookings scheduled",
+      "Least bookings scheduled",
+      "Most bookings completed",
+      "Least bookings completed",
+      "Most cancelled bookings",
+      "Most no-show (host)",
+      "Highest rated",
+      "Lowest rated",
+      "Recent ratings",
+      "Popular events",
+    ];
+
+    for (const title of expectedChartTitles) {
+      const chartCard = page.locator("[data-testid='panel-card'] h2").filter({ hasText: title });
+      await expect(chartCard).toBeVisible();
+    }
+  });
 });
